part 2/2 of fixing Altivec-accelerated H264 luma inloop filter
In h264_deblock_q1, the result of the deblock needs to be kept to
be used in future deblocks, so return this value now.

Also change the sign of tc0 vector: It is really a signed value, so
treat it as such until after the >=0 check;
then, at that point, after being masked, it can be treated as unsigned.

Patch by Graham Booker % gbooker A tamu P edu%


git-svn-id: svn://svn.mplayerhq.hu/ffmpeg/trunk@9349 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b